change some int to byte

+++ 00-LibrariesUsed/Tunes/Tunes.h
@@ -13,7 +13,7 @@ int superMarioTune[] = {
   NOTE_G6, NOTE_E7, NOTE_G7, NOTE_A7, 0, NOTE_F7, NOTE_G7, 0, NOTE_E7, 0,NOTE_C7, NOTE_D7, NOTE_B6, 0, 0
 };
 
-int superMarioTuneTempo[] = {
+byte superMarioTuneTempo[] = {
   12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12,12, 12, 12, 12, 
   12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 
   9, 9, 9, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12, 12,
@@ -21,7 +21,7 @@ int superMarioTuneTempo[] = {
   9, 9, 9, 12, 12, 12, 12, 12, 12, 12, 12,12, 12, 12, 12,
 };
 
-int superMarioTuneSize = sizeof(superMarioTune) / sizeof(int);
+byte superMarioTuneSize = sizeof(superMarioTune) / sizeof(int);
 
 /*******************************
  *  UNDERWORLD TUNE
@@ -34,13 +34,13 @@ int underworldTune[] = {
   NOTE_C4, NOTE_FS4,NOTE_F4, NOTE_E3, NOTE_AS4, NOTE_A4, NOTE_GS4, NOTE_DS4, NOTE_B3, NOTE_AS3, NOTE_A3, NOTE_GS3, 0, 0, 0
 };
 
-int underworldTuneTempo[] = { 
+byte underworldTuneTempo[] = { 
   12, 12, 12, 12,  12, 12, 6, 3, 12, 12, 12, 12,  12, 12, 6, 3, 12, 12, 12, 12, 12, 12, 6,3,
   12, 12, 12, 12, 12, 12, 6, 6, 18, 18, 18, 6, 6, 6, 6, 6, 6, 18, 18, 18, 18, 18, 18,
   10, 10, 10, 10, 10, 10, 3, 3, 3
 };
 
-int underworldTuneSize = sizeof(underworldTune) / sizeof(int);
+byte underworldTuneSize = sizeof(underworldTune) / sizeof(int);
 
 int starSpangledBannerTune[] = {
   NOTE_F5,NOTE_D5,NOTE_AS4,NOTE_D5,NOTE_F5,NOTE_AS5,NOTE_D6,NOTE_C6,NOTE_AS5,NOTE_D5,NOTE_E5,NOTE_F5,
@@ -60,7 +60,7 @@ float starSpangledBannerTuneTempo[] = {
   8,8,8/2.9,8,4,2,8,8,4,4,4,2, 4,4,4,8,8,4,4,4,4,8,8,8,8,4,4, 8,8,8/2.9,8,8,8,2,8,8,4,4,4,2
 };
 
-int starSpangledBannerTuneSize=sizeof(starSpangledBannerTune)/sizeof(int);
+byte starSpangledBannerTuneSize=sizeof(starSpangledBannerTune)/sizeof(int);
 
 
 /*******************************
@@ -78,7 +78,7 @@ float tuneTempo2A[] = {2, 3, 8, 2, 3, 16/3, 8, 8, 4, 4};
 float tuneTempo2B[] = {4, 2, 3, 16/3, 8, 8, 4, 3, 1000, 1000};
 float variantTuneTempo[] = {4, 2, 3, 8, 2, 8/3.0, 8, 3/2, 2, 1000};
 
-int tuneSize  = 10;
+byte tuneSize  = 10;
 
 /*******************************
  *  GAME OF THRONES 
@@ -95,8 +95,8 @@ int backTune2[] = {NOTE_C5, NOTE_DS3, NOTE_F3, NOTE_G3};
 int tunePeriod[] = {1175, 1175, 175, 175, 775, 775, 175, 175};
 int tunePeriodX[] = {1175, 1175, 575, 575, 575, 575};
 int backTune2Period[] = {375, 175, 175, 375};
-int backTuneSize = 4;
-int GOTTuneSize = 8;
-int tuneXSize = 6;
+byte backTuneSize = 4;
+byte GOTTuneSize = 8;
+byte tuneXSize = 6;
